Added documentation for the following changes
Added a flag to imd positions "-fold_chains" which allows chain molecules to be folded without losing bond connectivity.  A new function analyze_fold_chains was added to statistics_chain.c and this is now called by imd if the "-fold_chains" argument is specified.  Parsing for imd positions is now a separate function in imd.c called imd_parse_pos.  Also changed updateParticleCfg so that it now takes a flag argument: this flag specifies whether to get the bonding information or not.