This patch splits the virtual sites code into: * a general part * and implementations for specific rules for placing virtual particles
The existing routines, which place a virtual aprticle in the center of mass
of a molecule are now the "center of mass" implementation.

Also made several analysis commands conditional on the activation of the "center of mass implementation"

Adress requires the center of mass implementation. Added a compilation error, when adress is used without it.

Added an empty template for the "relative" implementation